Hybrid tuning algorithms
Hybrid tuning algorithms proposed in the article:
Giraldo, Sergio A. C., Príamo A. Melo, and Argimiro R. Secchi. 2022. "Tuning of Model Predictive Controllers Based on Hybrid optimization" Processes 10, no. 2:351. https://doi.org/10.3390/pr10020351
These codes were adapted to be used directly with the Matlab predictive control Toolbox, therefore the dynamic results may vary a little according to the answers shown in the Paper, however it follows the same methodology. This adaptation made it possible to distribute the code in a more agile way, allowing improvements and advances in the tuning branch of the MPC control.