MPCT update
The MPCObj object is now found as an output parameter of the MPCT function.

The minimum staggering has been added for linear systems where the pre- and post-multiplication matrices are located, which allows for improving the system's conditioning. All the internal calculation of the MPCT is done with the stepped linear model. Therefore, it is important that when implementing the MPC controller, the input and output signals received by the controller are staggered. To facilitate this, the MPCT function returns a ´scale´ variable that contains the conditioning matrices that can also be used for systems that present measured disturbances.

The MPCT now manages to improve the tuning for non-square processes where the number of controlled variables is greater than the number of manipulated variables, adding an initial trajectory of the impulse type as a reference, which is intended to be improved in the future.